+++ adapters/lora-vertx-base/src/test/java/org/eclipse/hono/adapter/lora/LoraProtocolAdapterTest.java
@@ -19,7 +19,9 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.RETURNS_SELF;
+import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.mockito.Mockito.withSettings;
@@ -29,6 +31,14 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
+import org.eclipse.hono.adapter.client.command.CommandConsumer;
+import org.eclipse.hono.adapter.client.command.CommandConsumerFactory;
+import org.eclipse.hono.adapter.client.command.CommandContext;
+import org.eclipse.hono.adapter.client.command.amqp.ProtonBasedCommand;
+import org.eclipse.hono.adapter.client.command.amqp.ProtonBasedCommandContext;
+import org.eclipse.hono.adapter.client.registry.TenantClient;
+import org.eclipse.hono.adapter.http.HttpAdapterMetrics;
+import org.eclipse.hono.adapter.lora.LoraCommand;
 import org.eclipse.hono.adapter.lora.LoraConstants;
 import org.eclipse.hono.adapter.lora.LoraMessage;
 import org.eclipse.hono.adapter.lora.LoraMessageType;
@@ -43,6 +53,8 @@ import org.eclipse.hono.client.ClientErrorException;
 import org.eclipse.hono.service.auth.DeviceUser;
 import org.eclipse.hono.service.http.HttpContext;
 import org.eclipse.hono.service.http.TracingHandler;
+import org.eclipse.hono.test.VertxMockSupport;
+import org.eclipse.hono.util.CommandEndpoint;
 import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.QoS;
 import org.eclipse.hono.util.RegistrationAssertion;
@@ -51,12 +63,20 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
+import io.micrometer.core.instrument.Timer.Sample;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.opentracing.Span;
 import io.opentracing.SpanContext;
 import io.opentracing.Tracer;
 import io.opentracing.Tracer.SpanBuilder;
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Context;
+import io.vertx.core.Future;
+import io.vertx.core.Handler;
+import io.vertx.core.Vertx;
 import io.vertx.core.buffer.Buffer;
+import io.vertx.core.http.HttpClient;
+import io.vertx.core.http.HttpClientRequest;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.http.HttpServerResponse;
 import io.vertx.core.json.JsonObject;
@@ -76,12 +96,18 @@ public class LoraProtocolAdapterTest extends ProtocolAdapterTestSupport<LoraProt
 
     private Tracer tracer;
     private Span currentSpan;
+    private Vertx vertx;
+    private Context context;
+    private HttpAdapterMetrics metrics;
 
     /**
      * Sets up the fixture.
      */
     @BeforeEach
     public void setUp() {
+
+        vertx = mock(Vertx.class);
+        context = VertxMockSupport.mockContext(vertx);
 
         this.properties = givenDefaultConfigurationProperties();
         createClientFactories();
@@ -96,9 +122,14 @@ public class LoraProtocolAdapterTest extends ProtocolAdapterTestSupport<LoraProt
         tracer = mock(Tracer.class);
         when(tracer.buildSpan(anyString())).thenReturn(spanBuilder);
 
+        metrics =  mock(HttpAdapterMetrics.class);
+        when(metrics.startTimer()).thenReturn(mock(Sample.class));
+
         adapter = new LoraProtocolAdapter();
         adapter.setConfig(properties);
         adapter.setTracer(tracer);
+        adapter.init(vertx, context);
+        adapter.setMetrics(metrics);
         setServiceClients(adapter);
     }
 
@@ -121,9 +152,14 @@ public class LoraProtocolAdapterTest extends ProtocolAdapterTestSupport<LoraProt
         final LoraProvider providerMock = getLoraProviderMock();
         final HttpContext httpContext = newHttpContext();
         final HttpServerRequest request = mock(HttpServerRequest.class);
+        final CommandConsumerFactory commandConsumerFactory = mock(CommandConsumerFactory.class);
         when(request.getHeader(eq(Constants.HEADER_QOS_LEVEL))).thenReturn(null);
         when(httpContext.request()).thenReturn(request);
+        final CommandConsumer commandConsumer = mock(CommandConsumer.class);
+        when(commandConsumer.close(any())).thenReturn(Future.succeededFuture());
+        when(commandConsumerFactory.createCommandConsumer(any(), any(), any(), any(), any())).thenReturn(Future.succeededFuture(commandConsumer));
 
+        adapter.setCommandConsumerFactory(commandConsumerFactory);
         adapter.handleProviderRoute(httpContext, providerMock);
 
         verify(httpContext.getRoutingContext()).put(LoraConstants.APP_PROPERTY_ORIG_LORA_PROVIDER, TEST_PROVIDER);
@@ -145,6 +181,111 @@ public class LoraProtocolAdapterTest extends ProtocolAdapterTestSupport<LoraProt
         assertThat(metaDataJson.getInteger(LoraConstants.APP_PROPERTY_FUNCTION_PORT)).isEqualTo(TEST_FUNCTION_PORT);
         verify(httpContext.getRoutingContext().response()).setStatusCode(HttpResponseStatus.ACCEPTED.code());
         verify(currentSpan).finish();
+    }
+
+    /**
+     * Verifies that an uplink message triggers a command subscription.
+     */
+    @Test
+    public void handleCommandSubscriptionSuccessfullyForUplinkMessage() {
+
+        givenATelemetrySenderForAnyTenant();
+
+        final LoraProvider providerMock = getLoraProviderMock();
+        final HttpContext httpContext = newHttpContext();
+        final HttpServerRequest request = mock(HttpServerRequest.class);
+        final CommandConsumerFactory commandConsumerFactory = mock(CommandConsumerFactory.class);
+        when(request.getHeader(eq(Constants.HEADER_QOS_LEVEL))).thenReturn(null);
+        when(httpContext.request()).thenReturn(request);
+        final CommandConsumer commandConsumer = mock(CommandConsumer.class);
+        when(commandConsumer.close(any())).thenReturn(Future.succeededFuture());
+        when(commandConsumerFactory.createCommandConsumer(any(), any(), any(), any(), any())).thenReturn(Future.succeededFuture(commandConsumer));
+
+        adapter.setCommandConsumerFactory(commandConsumerFactory);
+        adapter.handleProviderRoute(httpContext, providerMock);
+
+        verify(commandConsumerFactory).createCommandConsumer(eq("myTenant"), eq("myLoraGateway"),
+            VertxMockSupport.anyHandler(), eq(null), any());
+    }
+
+    /**
+     * Verifies that an uplink message triggers a command subscription.
+     *
+     */
+    @Test
+    public void handleCommandForLNS() {
+
+        givenATelemetrySenderForAnyTenant();
+
+        final LoraProvider providerMock = getLoraProviderMock();
+        final HttpContext httpContext = newHttpContext();
+        final HttpServerRequest request = mock(HttpServerRequest.class);
+        final CommandConsumerFactory commandConsumerFactory = mock(CommandConsumerFactory.class);
+        final TenantClient tenantClient = mock(TenantClient.class);
+        when(tenantClient.get(eq(TEST_TENANT_ID), any())).thenReturn(Future.succeededFuture(mock(TenantObject.class)));
+        when(request.getHeader(eq(Constants.HEADER_QOS_LEVEL))).thenReturn(null);
+        when(httpContext.request()).thenReturn(request);
+        final CommandConsumer commandConsumer = mock(CommandConsumer.class);
+        when(commandConsumer.close(any())).thenReturn(Future.succeededFuture());
+        when(commandConsumerFactory.createCommandConsumer(any(), any(), any(), any(), any())).thenReturn(Future.succeededFuture(commandConsumer));
+
+        adapter.setCommandConsumerFactory(commandConsumerFactory);
+        adapter.handleProviderRoute(httpContext, providerMock);
+        adapter.setTenantClient(tenantClient);
+
+        final ArgumentCaptor<Handler<CommandContext>> handlerArgumentCaptor = VertxMockSupport.argumentCaptorHandler();
+
+        verify(commandConsumerFactory).createCommandConsumer(eq(TEST_TENANT_ID),
+            eq(TEST_GATEWAY_ID), handlerArgumentCaptor.capture(), eq(null), any());
+
+        final Handler<CommandContext> commandHandler = handlerArgumentCaptor.getValue();
+        final ProtonBasedCommand command = mock(ProtonBasedCommand.class);
+        when(command.getTenant()).thenReturn(TEST_TENANT_ID);
+        when(command.getDeviceId()).thenReturn(TEST_DEVICE_ID);
+        when(command.getGatewayId()).thenReturn(TEST_GATEWAY_ID);
+        when(command.getPayload()).thenReturn(Buffer.buffer("bumlux".getBytes(StandardCharsets.UTF_8)));
+        when(command.isValid()).thenReturn(true);
+
+        final ProtonBasedCommandContext commandContext = mock(ProtonBasedCommandContext.class);
+        when(commandContext.getCommand()).thenReturn(command);
+        when(commandContext.getTracingSpan()).thenReturn(currentSpan);
+
+        final RegistrationAssertion gatewayRegistration = new RegistrationAssertion(TEST_GATEWAY_ID);
+        final CommandEndpoint commandEndpoint = new CommandEndpoint();
+        commandEndpoint.setHeaders(Map.of("my-header", "my-header-value"));
+        commandEndpoint.setUri("https://my-server.com/commands/{{deviceId}}/send");
+        gatewayRegistration.setCommandEndpoint(commandEndpoint);
+
+        final JsonObject json = new JsonObject();
+        json.put("my-payload", "bumlux");
+        final LoraCommand loraCommand = new LoraCommand(json, "https://my-server.com/commands/deviceId/send");
+        when(providerMock.getCommand(any(), any(), any())).thenReturn(loraCommand);
+        when(providerMock.getDefaultHeaders()).thenReturn(Map.of("my-provider-header", "my-provider-header-value"));
+
+        final HttpClient httpClient = mock(HttpClient.class);
+        when(vertx.createHttpClient()).thenReturn(httpClient);
+
+        final HttpClientRequest httpClientRequest = mock(HttpClientRequest.class);
+        when(httpClient.postAbs(anyString())).thenReturn(httpClientRequest);
+        when(httpClientRequest.handler(any())).thenReturn(httpClientRequest);
+        when(httpClientRequest.exceptionHandler(any())).thenReturn(httpClientRequest);
+
+        doAnswer(invocation -> {
+            final Handler<AsyncResult> callback = invocation.getArgument(1);
+            callback.handle(Future.succeededFuture());
+            return null;
+        }).when(httpClientRequest).write(anyString(), any(Handler.class));
+
+        when(registrationClient.assertRegistration(eq(LoraProtocolAdapterTest.TEST_TENANT_ID),
+            eq(LoraProtocolAdapterTest.TEST_GATEWAY_ID), eq(null), any())).thenReturn(Future.succeededFuture(gatewayRegistration));
+
+        commandHandler.handle(commandContext);
+
+        verify(vertx, times(1)).createHttpClient();
+        verify(httpClient, times(1)).postAbs("https://my-server.com/commands/deviceId/send");
+        verify(httpClientRequest, times(1)).putHeader("my-header", "my-header-value");
+        verify(httpClientRequest, times(1)).putHeader("my-provider-header", "my-provider-header-value");
+        verify(httpClientRequest, times(1)).end(eq(json.encode()), any(Handler.class));
     }
 
     /**