+++ adapters/lora-vertx-base/src/test/java/org/eclipse/hono/adapter/lora/providers/LoriotProviderTest.java
@@ -18,6 +18,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.eclipse.hono.adapter.lora.LoraMetaData;
 import org.eclipse.hono.adapter.lora.UplinkLoraMessage;
 
+import io.vertx.core.json.JsonObject;
+
 /**
  * Verifies behavior of {@link LoriotProvider}.
  */
@@ -50,5 +52,20 @@ public class LoriotProviderTest extends LoraProviderTestBase<LoriotProvider> {
         assertThat(metaData.getGatewayInfo().get(0).getSnr()).isEqualTo(10);
         assertThat(metaData.getGatewayInfo().get(0).getLocation().getLongitude()).isEqualTo(4.4007817);
         assertThat(metaData.getGatewayInfo().get(0).getLocation().getLatitude()).isEqualTo(12);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected void assertCommandFormat(final JsonObject command) {
+        assertThat(command.containsKey("cmd")).isTrue();
+        assertThat(command.getValue("cmd")).isEqualTo("tx");
+        assertThat(command.containsKey("confirmed")).isTrue();
+        assertThat((boolean) command.getValue("confirmed")).isFalse();
+        assertThat(command.containsKey("data")).isTrue();
+        assertThat(command.getValue("data")).isEqualTo("62756D6C7578");
+        assertThat(command.containsKey("EUI")).isTrue();
+        assertThat(command.getValue("EUI")).isEqualTo(TEST_DEVICE_ID);
     }
 }